refactor(request_service): relocate BlockedSurface into RequestService namespace
